requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
pymongo==4.3.3
motor==3.1.1
python-dotenv==1.0.0
aiohttp==3.8.4
dnspython==2.3.0
urllib3==1.26.15
